# Conflicts:
#	composeApp/src/commonMain/composeResources/values-ar/strings.xml
#	composeApp/src/commonMain/composeResources/values/strings.xml

+++ composeApp/src/commonMain/kotlin/org/ninjaneers/motager/dashboard/presentation/customers/data/dto/CustomerDTO.kt
@@ -13,5 +13,5 @@ data class CustomerDTO(
     @SerialName("number_of_orders")
     val numberOfOrders: Int?,
     @SerialName("total_spent")
-    val totalSpent: Int?
+    val totalSpent: Double?
 )
